+++ src/predicate_lang.ml
@@ -68,9 +68,21 @@ module Ast = struct
     match kind with
     | Dune -> many union [] kind
     | Jbuild -> one kind
+
+  let rec to_sexp f =
+    let open Sexp.Encoder in
+    function
+    | Element a -> f a
+    | Compl a -> constr "compl" [to_sexp f a]
+    | Standard -> string ":standard"
+    | Union xs -> constr "or" (List.map ~f:(to_sexp f) xs)
+    | Inter xs -> constr "and" (List.map ~f:(to_sexp f) xs)
 end
 
 type t = (string -> bool) Ast.t
+
+let pp ppf t =
+  Sexp.pp ppf (Ast.to_sexp (fun _ -> Sexp.Encoder.string "opaque") t)
 
 let decode : t Dune_lang.Decoder.t =
   let open Stanza.Decoder in
@@ -88,14 +100,16 @@ let rec mem t ~standard ~elem =
 
 let filter (t : t) ~standard elems =
   match t with
+  | Inter []
   | Union [] -> []
   | _ ->
-    (List.filter (Lazy.force elems)
-       ~f:(fun elem -> mem t ~standard:standard ~elem))
+    (List.filter elems ~f:(fun elem -> mem t ~standard ~elem))
 
 let union t = Ast.Union t
 
 let of_glob g = Ast.Element (Glob.test g)
+
+let of_pred p = Ast.Element p
 
 let of_string_set s = Ast.Element (String.Set.mem s)